Move pregel rest handling into pregel directory
