+++ arangod/Pregel/CMakeLists.txt
@@ -13,9 +13,7 @@ add_library(arango_pregel STATIC
   Reports.h
   Utils.cpp
   Worker.cpp
-  WorkerConfig.cpp
-  ${PROJECT_SOURCE_DIR}/arangod/RestHandler/RestControlPregelHandler.cpp
-  ${PROJECT_SOURCE_DIR}/arangod/RestHandler/RestPregelHandler.cpp)
+  WorkerConfig.cpp)
 
 target_link_libraries(arango_pregel
 	arango
@@ -28,3 +26,4 @@ target_include_directories(arango_pregel PRIVATE
   "${PROJECT_SOURCE_DIR}/${ENTERPRISE_INCLUDE_DIR}")
 
 add_subdirectory(Algos)
+add_subdirectory(REST)
