+++ arangod/Pregel/REST/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(arango_pregel PRIVATE
+  RestControlPregelHandler.cpp
+  RestPregelHandler.cpp)
